Pcentra/requirements.txt

asgiref==3.4.1
Django==3.2.9
djangorestframework==3.11.0
pytz==2021.3
sqlparse==0.4.2
